FileDialog: uses SINGLE_THREADED ThreadPool on all Windows targets to prevent application hang (closes #1946) (references #1849)
Previously, it seemed that only Windows HashLink would hang when opening a FileDialog, but I was able to reproduce #1946 on Windows CPP too. Not sure why I couldn't before because I don't see any obvious changes to FileDialog or ThreadPool that would cause it to happen now, but not when it was affecting HashLink previously.

Does not seem to affect other operating systems.

We might consider going back to BackgroundWorker instead, which we used in Lime 8.1 and older.